feat: addressing copilot duplicated service declaration
